pyproject.toml
[tool.poetry]
name = "shitcoding_gendiff"
version = "0.0.5"
description = "Shitcoding project #2: Gendiff script: check differences between two files"
authors = ["Shitcoder"]
packages = [
	{ include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.6"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
pytest = "^6.0.1"
pytest-cov = "^2.10.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

